Feat: Move filter classes to a separate file
